backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==10.4.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.31.0
setuptools==80.9.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
huggingface_hub==0.20.0
pytesseract==0.3.10

